requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
backports.zoneinfo==0.2.1
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
dataclasses-json==0.6.4
distro==1.9.0
Django==4.2.10
exceptiongroup==1.2.0
frozenlist==1.4.1
greenlet==3.0.3
grpcio==1.62.0
grpcio-tools==1.62.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.4
httpx==0.27.0
hyperframe==6.0.1
idna==3.6
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.9
langchain-community==0.0.24
langchain-core==0.1.26
langchain-openai==0.0.7
langsmith==0.1.8
lxml==5.1.0
marshmallow==3.20.2
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.24.4
openai==1.12.0
orjson==3.9.15
packaging==23.2
portalocker==2.8.2
protobuf==4.25.3
pydantic==2.6.2
pydantic-core==2.16.3
python-docx==1.1.0
PyYAML==6.0.1
qdrant-client==1.7.3
regex==2023.12.25
requests==2.31.0
sniffio==1.3.1
SQLAlchemy==2.0.27
sqlparse==0.4.4
tenacity==8.2.3
tiktoken==0.6.0
tqdm==4.66.2
typing-extensions==4.10.0
typing-inspect==0.9.0
tzdata==2024.1
urllib3==2.2.1
yarl==1.9.4